moved ParseNameFromHoverText to Utils; moved AdjustForTileSize & DrawGiftTooltip to GiftHelper for use in SocialPage; created DrawGiftCalTooltip
