+++ GiftTasteHelper/CalendarGiftHelper.cs
@@ -77,7 +77,7 @@ namespace GiftTasteHelper
                 {
                     Utils.DebugLog("hover text: " + hoverText);
 
-                    string npcName = Calendar.ParseNameFromHoverText(hoverText);
+                    string npcName = Utils.ParseNameFromHoverText(hoverText);
                     Debug.Assert(npcGiftInfo.ContainsKey(npcName));
 
                     currentGiftInfo = npcGiftInfo[npcName];
@@ -100,11 +100,11 @@ namespace GiftTasteHelper
             // Double check here since we may not be unsubscribed from post render right away when the calendar closes
             if (drawCurrentFrame && currentGiftInfo != null)
             {
-                DrawGiftTooltip(currentGiftInfo, "Favourite Gifts");
+                DrawGiftCalTooltip(currentGiftInfo, "Favourite Gifts");
             }
         }
 
-        private void DrawGiftTooltip(NPCGiftInfo giftInfo, string title)
+        private void DrawGiftCalTooltip(NPCGiftInfo giftInfo, string title)
         {
             int numItems = giftInfo.FavouriteGifts.Length;
             if (numItems == 0)
@@ -130,60 +130,15 @@ namespace GiftTasteHelper
             int viewportW = Game1.viewport.Width;
             int viewportH = Game1.viewport.Height;
 
-            // Reduce the number items shown if it will go off screen.
-            // TODO: add a scrollbar or second column
-            if (height > viewportH)
-            {
-                numItems = (viewportH / rowHeight) - 1; // Remove an item to make space for the title
-                height = AdjustForTileSize(rowHeight * numItems);
-            }
-
             // Approximate where the original tooltip will be positioned
             SVector2 origHoverTextSize = SVector2.MeasureString(calendar.GetCurrentHoverText(), Game1.dialogueFont);
-            int origTToffsetX = Math.Max(0, AdjustForTileSize(origHoverTextSize.x + mouse.x, 1.0f) - viewportW);
+            int origTToffsetX = Math.Max(0, GiftHelper.AdjustForTileSize(origHoverTextSize.x + mouse.x, 1.0f) - viewportW);
 
             // Consider the position of the original tooltip and ensure we don't cover it up
             SVector2 tooltipPos = ClampToViewport(x - origTToffsetX, y, width, height, viewportW, viewportH);
 
-            // Draw the background of the tooltip
-            SpriteBatch spriteBatch = Game1.spriteBatch;
-#if SMAPI_VERSION_39_3_AND_PRIOR
-            spriteBatch.Begin(SpriteSortMode.Immediate, BlendState.AlphaBlend, SamplerState.PointClamp, null, null);
-#endif
-            Rectangle menuTextureSourceRect = new Rectangle(0, 256, 60, 60);
-            IClickableMenu.drawTextureBox(spriteBatch, Game1.menuTexture, menuTextureSourceRect, tooltipPos.xi, tooltipPos.yi, width, height, Color.White, ZoomLevel);
-
-            // Offset the sprite from the corner of the bg, and the text to the right and centered vertically of the sprite
-            SVector2 spriteOffset = new SVector2(AdjustForTileSize(tooltipPos.x, 0.25f), AdjustForTileSize(tooltipPos.y, 0.25f));
-            SVector2 textOffset = new SVector2(spriteOffset.x, spriteOffset.y + (spriteRect.Height / 2));
-
-            // Draw the title then set up the offset for the remaining text
-            DrawText(title, textOffset);
-            textOffset.x += scaledSpriteSize.x + padding;
-            textOffset.y += rowHeight;
-            spriteOffset.y += rowHeight;
-
-            for (int i = 0; i < numItems; ++i)
-            {
-                NPCGiftInfo.ItemData item = giftInfo.FavouriteGifts[i];
-
-                // Draw the sprite for the item then the item text
-                DrawText(item.name, textOffset);
-                DrawTexture(Game1.objectSpriteSheet, spriteOffset, item.tileSheetSourceRect, spriteScale);
-
-                // Move to the next row
-                spriteOffset.y += rowHeight;
-                textOffset.y += rowHeight;
-            }
-#if SMAPI_VERSION_39_3_AND_PRIOR
-            spriteBatch.End();
-#endif
-        }
-
-        private int AdjustForTileSize(float v, float tileSizeMod = 0.5f, float zoom = 1.0f)
-        {
-            float tileSize = (float)Game1.tileSize * tileSizeMod;
-            return (int)((v + tileSize) * zoom);
+            // Draw the tooltip
+            DrawGiftTooltip(giftInfo, title, tooltipPos);
         }
 
         private SVector2 ClampToViewport(int x, int y, int w, int h, int viewportW, int viewportH)
