+++ GiftTasteHelper/GiftHelper.cs
@@ -99,6 +99,11 @@ namespace GiftTasteHelper
         {
             // Empty
         }
+        public static int AdjustForTileSize(float v, float tileSizeMod = 0.5f, float zoom = 1.0f)
+        {
+            float tileSize = (float)Game1.tileSize * tileSizeMod;
+            return (int)((v + tileSize) * zoom);
+        }
 
         protected void DrawText(string text, SVector2 pos)
         {
@@ -116,5 +121,77 @@ namespace GiftTasteHelper
 #endif
             Game1.spriteBatch.Draw(texture, pos.ToXNAVector2(), source, Color.White, 0.0f, Vector2.Zero, scale, SpriteEffects.None, 0.0f);
         }
+
+
+        public void DrawGiftTooltip(NPCGiftInfo giftInfo, string title, SVector2 tooltipPos)
+        {
+            int numItems = giftInfo.FavouriteGifts.Length;
+            if (numItems == 0)
+                return;
+
+            float spriteScale = 2.0f * ZoomLevel; // 16x16 is pretty small
+            Rectangle spriteRect = giftInfo.FavouriteGifts[0].tileSheetSourceRect; // We just need the dimensions which we assume are all the same
+            SVector2 scaledSpriteSize = new SVector2(spriteRect.Width * spriteScale, spriteRect.Height * spriteScale);
+
+            // The longest length of text will help us determine how wide the tooltip box should be 
+            SVector2 titleSize = SVector2.MeasureString(title, Game1.smallFont);
+            SVector2 maxTextSize = (titleSize.x - scaledSpriteSize.x > giftInfo.MaxGiftNameSize.x) ? titleSize : giftInfo.MaxGiftNameSize;
+
+            SVector2 mouse = new SVector2(Game1.getOldMouseX(), Game1.getOldMouseY());
+
+            int padding = 4;
+            int rowHeight = (int)Math.Max(maxTextSize.y * ZoomLevel, scaledSpriteSize.yi) + padding;
+            int width = AdjustForTileSize((maxTextSize.x * ZoomLevel) + scaledSpriteSize.xi) + padding;
+            int height = AdjustForTileSize(rowHeight * (numItems + 1), 0.5f); // Add one to make room for the title
+            int x = AdjustForTileSize(mouse.x, 0.5f, ZoomLevel) - width;
+            int y = AdjustForTileSize(mouse.y, 0.5f, ZoomLevel);
+
+            int viewportW = Game1.viewport.Width;
+            int viewportH = Game1.viewport.Height;
+
+            // Reduce the number items shown if it will go off screen.
+            // TODO: add a scrollbar or second column
+            if (height > viewportH)
+            {
+                numItems = (viewportH / rowHeight) - 1; // Remove an item to make space for the title
+                height = AdjustForTileSize(rowHeight * numItems);
+            }
+
+            // Draw the background of the tooltip
+            SpriteBatch spriteBatch = Game1.spriteBatch;
+#if SMAPI_VERSION_39_3_AND_PRIOR
+            spriteBatch.Begin(SpriteSortMode.Immediate, BlendState.AlphaBlend, SamplerState.PointClamp, null, null);
+#endif
+            Rectangle menuTextureSourceRect = new Rectangle(0, 256, 60, 60);
+            IClickableMenu.drawTextureBox(spriteBatch, Game1.menuTexture, menuTextureSourceRect, tooltipPos.xi, tooltipPos.yi, width, height, Color.White, ZoomLevel);
+
+            // Offset the sprite from the corner of the bg, and the text to the right and centered vertically of the sprite
+            SVector2 spriteOffset = new SVector2(AdjustForTileSize(tooltipPos.x, 0.25f), AdjustForTileSize(tooltipPos.y, 0.25f));
+            SVector2 textOffset = new SVector2(spriteOffset.x, spriteOffset.y + (spriteRect.Height / 2));
+
+            // Draw the title then set up the offset for the remaining text
+            DrawText(title, textOffset);
+            textOffset.x += scaledSpriteSize.x + padding;
+            textOffset.y += rowHeight;
+            spriteOffset.y += rowHeight;
+
+            for (int i = 0; i < numItems; ++i)
+            {
+                NPCGiftInfo.ItemData item = giftInfo.FavouriteGifts[i];
+
+                // Draw the sprite for the item then the item text
+                DrawText(item.name, textOffset);
+                DrawTexture(Game1.objectSpriteSheet, spriteOffset, item.tileSheetSourceRect, spriteScale);
+
+                // Move to the next row
+                spriteOffset.y += rowHeight;
+                textOffset.y += rowHeight;
+            }
+#if SMAPI_VERSION_39_3_AND_PRIOR
+            spriteBatch.End();
+#endif
+        }
     }
+
 }
+
