+++ GiftTasteHelper/GiftTasteHelper.sln
@@ -1,0 +1,22 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 14
+VisualStudioVersion = 14.0.25123.0
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "GiftTasteHelper-d", "GiftTasteHelper-d.csproj", "{CA415B12-5C37-487A-92BE-0AED14440F36}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CA415B12-5C37-487A-92BE-0AED14440F36}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA415B12-5C37-487A-92BE-0AED14440F36}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA415B12-5C37-487A-92BE-0AED14440F36}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA415B12-5C37-487A-92BE-0AED14440F36}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+EndGlobal
