+++ GiftTasteHelper/manifest.json
@@ -1,0 +1,14 @@
+﻿{
+  "Name": "GiftTasteHelper-d",
+  "Authour": "dreamsicl",
+  "Version": {
+    "MajorVersion": 1,
+    "MinorVersion": 1,
+    "PatchVersion": 1,
+    "Build": ""
+  },
+  "Description": "Displays NPC gift tastes in a handy tooltip.",
+  "UniqueID": "",
+  "PerSaveConfigs": false,
+  "EntryDll": "GiftTasteHelper-d.dll"
+}
